requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astpretty==3.0.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
configmypy==0.2.0
contourpy==1.3.2
crc32c==2.7.1
cuda-bindings==12.9.2
cuda-core==0.3.2
cuda-pathfinder==1.2.2
cuda-quantum-cu12==0.10.0
cudaq==0.10.0
cudensitymat-cu12==0.3.0
cupy-cuda12x==13.4.1
cuquantum-python-cu12==25.9.0
custatevec-cu12==1.10.0
cutensor-cu12==2.3.1
cutensornet-cu12==2.9.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
donfig==0.8.1.post1
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.3.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook_shim==0.2.4
numcodecs==0.16.0
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvmath-python==0.6.0
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pipreqs==0.4.13
platformdirs==4.3.8
pluggy==1.5.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.3
Send2Trash==1.8.3
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tensorly==0.9.0
tensorly-torch==0.5.0
terminado==0.18.1
tinycss2==1.4.0
torch==2.7.0
torch_harmonics==0.7.4
tornado==6.5.1
traitlets==5.14.3
triton==3.3.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
yarg==0.1.10
zarr==3.0.7
neuraloperator==2.0.0